Restructure of code to fit best practice package structure
