+++ apps/create_sim.py
@@ -2,7 +2,7 @@ import dash_core_components as dcc
 import dash_html_components as html
 from dash.dependencies import Input, Output
 
-from src.frontend.app import app
+from app import app
 
 # TODO: Add/remove users from the neighbourhood
 # TODO: Number of days simulated
