+++ backend/optimizer.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from scipy import optimize, math
 
-from src import load_profile_utils
+from util import load_profile_utils
 
 
 class Optimizer:
