--- src/util/message_utils.py
@@ -1,17 +1,0 @@
-from enum import Enum
-
-from src.backend.job import Job
-
-
-class Action(Enum):
-    request = 'REQUEST'
-    broadcast = 'BROADCAST'
-    cancel = 'CANCEL'
-    accept = 'ACCEPT'
-    decline = 'DECLINE'
-    prediction = 'PREDICTION'
-
-
-def job_from_message(message):
-    return Job(est=message['est'], lst=message['lst'],
-               load_profile=message['load_profile'])
